[XFS] Fix warnings when tracing enabled on 64 bit platforms
SGI Modid: 2.5.x-xfs:slinx:160622a

+++ fs/xfs/xfs_buf_item.c
@@ -1214,7 +1214,7 @@ xfs_buf_item_trace(
 		     (void *)((unsigned long)XFS_BUF_BFLAGS(bp)),
 		     XFS_BUF_FSPRIVATE(bp, void *),
 		     XFS_BUF_FSPRIVATE2(bp, void *),
-		     (void *)XFS_BUF_ISPINNED(bp),
+		     (void *)(unsigned long)XFS_BUF_ISPINNED(bp),
 		     (void *)XFS_BUF_IODONE_FUNC(bp),
 		     (void *)((unsigned long)(XFS_BUF_VALUSEMA(bp))),
 		     (void *)bip->bli_item.li_desc,
